requirement.txt
azure-ai-projects>=2.0.0b1
azure-identity
openai~=2.0.1